Created test scripts to learn basics
Created multiple python programs that use the RLGym frame work to create a machine learned model that controls a bot.

First I created a the most simple script I could that would teach a bot using PPO: FirstPPO_rlgym

Second I created a script that specified custom rewards to give the algorithm: BasicRewards_rlgym

Last I created a script that created a bot that learns in a 1v1 environment with custom rewards. This allows me to play against the bot using RLBot.

A duplicate of the ExampleBot provided by RLGym has been created and updated with the settings and model that my scripts taught.